requirements.txt
esda>=2.4.1
geopandas>=0.10.2
matplotlib>=3.1.2
matplotlib_scalebar>=0.8.1
numpy>=1.17.4
osmnx>=1.4.0
pandas>=1.3.4
Pyomo>=6.2
PyYAML>=6.0.1
scikit_learn>=1.3.2
Shapely>=1.7.0
tqdm>=4.62.3
pyarrow>=13.0.0
